Remove out_normals from transform_SSD/transform_dual_quat; the data is never used.
Clarify that the second parameter of smooth_mesh isn't necessarily normals;
it's really just a scratch buffer.

Rename full_eval to final_pass for clarity.

Fix some 64-bit warnings.

Remove unused d_ssd_normals, d_packed_vert_map.
Remove Mesh._tangents and quads.

+++ include/animation/graph.hpp
@@ -92,10 +92,10 @@ struct Graph {
   void set_offset_scale(const Vec3_cu& _offset, float _scale);
 
   /// Return the number of vertices
-  inline int nb_vertices() const { return _vertices.size(); }
+  inline size_t nb_vertices() const { return _vertices.size(); }
 
   /// Return the number of edges
-  inline int nb_edges() const { return _edges.size(); }
+  inline size_t nb_edges() const { return _edges.size(); }
 
   //----------------------------------------------------------------------------
   /// @name Attributes
